# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardSummaryThroughput = {
  current: number;
  target: number;
  unit: string;
};

export interface DashboardSummary {
  receivedToday: number;
  processedToday: number;
  quarantineCount: number;
  expiringSoonCount: number;
  verifiedRate: number;
  throughput: DashboardSummaryThroughput;
}

export type UnitStatus = typeof UnitStatus[keyof typeof UnitStatus];


export const UnitStatus = {
  pending: 'pending',
  verified: 'verified',
  quarantine: 'quarantine',
  processed: 'processed',
  disposed: 'disposed',
} as const;

export type UnitScreeningGate = typeof UnitScreeningGate[keyof typeof UnitScreeningGate];


export const UnitScreeningGate = {
  awaiting: 'awaiting',
  passed: 'passed',
  quarantine: 'quarantine',
} as const;

export type UnitScreening = {
  aboForward: string;
  aboReverse: string;
  rhd: string;
  hiv: string;
  hbsag: string;
  hcv: string;
  vdrl: string;
  gate: UnitScreeningGate;
};

export interface Unit {
  id: string;
  donorCode: string;
  collectionSite: string;
  collectedAt: string;
  volumeMl: number;
  /** @nullable */
  bloodGroup: string | null;
  status: UnitStatus;
  screening: UnitScreening;
  createdAt: string;
}

export interface UnitInput {
  /** @minLength 3 */
  donorCode: string;
  /** @minLength 2 */
  collectionSite: string;
  collectedAt: string;
  /** @minimum 1 */
  volumeMl: number;
}

export type SerologyInputHiv = typeof SerologyInputHiv[keyof typeof SerologyInputHiv];


export const SerologyInputHiv = {
  'non-reactive': 'non-reactive',
  reactive: 'reactive',
} as const;

export type SerologyInputHbsag = typeof SerologyInputHbsag[keyof typeof SerologyInputHbsag];


export const SerologyInputHbsag = {
  'non-reactive': 'non-reactive',
  reactive: 'reactive',
} as const;

export type SerologyInputHcv = typeof SerologyInputHcv[keyof typeof SerologyInputHcv];


export const SerologyInputHcv = {
  'non-reactive': 'non-reactive',
  reactive: 'reactive',
} as const;

export type SerologyInputVdrl = typeof SerologyInputVdrl[keyof typeof SerologyInputVdrl];


export const SerologyInputVdrl = {
  'non-reactive': 'non-reactive',
  reactive: 'reactive',
} as const;

export interface SerologyInput {
  aboForward: string;
  aboReverse: string;
  rhd: string;
  hiv: SerologyInputHiv;
  hbsag: SerologyInputHbsag;
  hcv: SerologyInputHcv;
  vdrl: SerologyInputVdrl;
}

export interface FractionateInput {
  operator?: string;
  shelf?: string;
}

export type ComponentType = typeof ComponentType[keyof typeof ComponentType];


export const ComponentType = {
  PRBC: 'PRBC',
  FFP: 'FFP',
  PLT: 'PLT',
} as const;

export type ComponentStatus = typeof ComponentStatus[keyof typeof ComponentStatus];


export const ComponentStatus = {
  available: 'available',
  expiring: 'expiring',
  dispatched: 'dispatched',
  expired: 'expired',
} as const;

export interface Component {
  id: string;
  parentUnitId: string;
  type: ComponentType;
  volumeMl: number;
  storage: string;
  expiresAt: string;
  shelf: string;
  status: ComponentStatus;
}

/**
 * @nullable
 */
export type VaultSlotComponent = {
  id?: string;
  type?: string;
  bloodGroup?: string;
  expiresAt?: string;
  hoursRemaining?: number;
  parentUnitId?: string;
} | null;

export type VaultSlotStatus = typeof VaultSlotStatus[keyof typeof VaultSlotStatus];


export const VaultSlotStatus = {
  occupied: 'occupied',
  empty: 'empty',
  expiring: 'expiring',
  quarantine: 'quarantine',
} as const;

export interface VaultSlot {
  shelf: string;
  zone: string;
  temperature: string;
  /** @nullable */
  component: VaultSlotComponent;
  status: VaultSlotStatus;
}

export type FeedbackInputPlatform = typeof FeedbackInputPlatform[keyof typeof FeedbackInputPlatform];


export const FeedbackInputPlatform = {
  'crucible-lab': 'crucible-lab',
} as const;

export type FeedbackInputPaperComparison = typeof FeedbackInputPaperComparison[keyof typeof FeedbackInputPaperComparison];


export const FeedbackInputPaperComparison = {
  'much-slower': 'much-slower',
  slower: 'slower',
  same: 'same',
  faster: 'faster',
  'much-faster': 'much-faster',
} as const;

export interface FeedbackInput {
  platform: FeedbackInputPlatform;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  paperComparison: FeedbackInputPaperComparison;
  note?: string;
}

export type Feedback = FeedbackInput & {
  id: string;
  submittedAt: string;
};

export type GetUnitsParams = {
status?: GetUnitsStatus;
search?: string;
};

export type GetUnitsStatus = typeof GetUnitsStatus[keyof typeof GetUnitsStatus];


export const GetUnitsStatus = {
  all: 'all',
  pending: 'pending',
  verified: 'verified',
  quarantine: 'quarantine',
  processed: 'processed',
  disposed: 'disposed',
} as const;