# packages/api-zod/src/generated/types/component.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ComponentStatus } from './componentStatus';
import type { ComponentType } from './componentType';

export interface Component {
  id: string;
  parentUnitId: string;
  type: ComponentType;
  volumeMl: number;
  storage: string;
  expiresAt: string;
  shelf: string;
  status: ComponentStatus;
}