# packages/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryThroughput } from './dashboardSummaryThroughput';

export interface DashboardSummary {
  receivedToday: number;
  processedToday: number;
  quarantineCount: number;
  expiringSoonCount: number;
  verifiedRate: number;
  throughput: DashboardSummaryThroughput;
}