# packages/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInput } from './feedbackInput';

export type Feedback = FeedbackInput & {
  id: string;
  submittedAt: string;
};