# packages/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputPaperComparison } from './feedbackInputPaperComparison';
import type { FeedbackInputPlatform } from './feedbackInputPlatform';

export interface FeedbackInput {
  platform: FeedbackInputPlatform;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  paperComparison: FeedbackInputPaperComparison;
  note?: string;
}