# packages/api-zod/src/generated/types/serologyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SerologyInputHbsag } from './serologyInputHbsag';
import type { SerologyInputHcv } from './serologyInputHcv';
import type { SerologyInputHiv } from './serologyInputHiv';
import type { SerologyInputVdrl } from './serologyInputVdrl';

export interface SerologyInput {
  aboForward: string;
  aboReverse: string;
  rhd: string;
  hiv: SerologyInputHiv;
  hbsag: SerologyInputHbsag;
  hcv: SerologyInputHcv;
  vdrl: SerologyInputVdrl;
}