# packages/api-zod/src/generated/types/unit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UnitScreening } from './unitScreening';
import type { UnitStatus } from './unitStatus';

export interface Unit {
  id: string;
  donorCode: string;
  collectionSite: string;
  collectedAt: string;
  volumeMl: number;
  /** @nullable */
  bloodGroup: string | null;
  status: UnitStatus;
  screening: UnitScreening;
  createdAt: string;
}