# packages/api-zod/src/generated/types/vaultSlot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VaultSlotComponent } from './vaultSlotComponent';
import type { VaultSlotStatus } from './vaultSlotStatus';

export interface VaultSlot {
  shelf: string;
  zone: string;
  temperature: string;
  /** @nullable */
  component: VaultSlotComponent;
  status: VaultSlotStatus;
}